Implement feature in colors to apply a colorbar to discrete data Fixes #17
Adds rgba_to_hex function to convert rgba colors to hex form, since the colors library is all in hex form which I prefer.
Adds get_discrete_colormap, which forms an array of colors for a set of parameters that are scaled in color according to a given matplotlib color map.